+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/996>)
 - Add VocInstanceSegmentationImporter and VocInstanceSegmentationExporter
   (<https://github.com/openvinotoolkit/datumaro/pull/997>)
+- Add Segment Anything data format support
+  (<https://github.com/openvinotoolkit/datumaro/pull/1005>)
 
 ### Enhancements
 - Use autosummary for fully-automatic Python module docs generation